GH-13: Added the eslint-plugin-react, and refactored code in regards to the new rules.
