+++ src/Components/Button/Button.stories.tsx
@@ -16,7 +16,6 @@ export const HelloWorld = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 HelloWorld.args = {
   label: 'Hello world!',
-  
 };
 
 export const ClickMe = Template.bind({});
