+++ src/Components/Input/index.ts
@@ -1,2 +1,1 @@
 export * from './BaseInput';
-
