+++ src/Components/Modal/Modal.stories.tsx
@@ -5,73 +5,68 @@ import { Alignment } from 'index';
 import React from 'react';
 import { Modal } from './Modal';
 
-
-
-
 export default {
-    title: 'Egde Library/Modal',
-    component: Modal,
-    parameters: {
-        layout: 'centered'
+  title: 'Egde Library/Modal',
+  component: Modal,
+  parameters: {
+    layout: 'centered',
+  },
+  argTypes: {
+    isModalOpen: { defaultValue: false },
+    hasCloseButton: { defaultValue: true },
+    buttons: {
+      variant: {
+        primary: {
+          label: '',
+          variant: '',
+          onClick: undefined,
+        },
+        secondary: {
+          label: '',
+          variant: '',
+          onClick: undefined,
+        },
+        alignment: Alignment.Left,
+      },
     },
-    argTypes: {
-        isModalOpen: {defaultValue: false},
-        hasCloseButton: {defaultValue: true}, 
-        buttons: {variant: {
-                primary: {
-                    label: '',
-                    variant: '',
-                    onClick: undefined
-                },
-                secondary: {
-                    label: '',
-                    variant: '',
-                    onClick: undefined
-                },
-                alignment: Alignment.Left
-            }}
-    }
+  },
 } as ComponentMeta<typeof Modal>;
 
-const Template: ComponentStory<typeof Modal> = (args) => <Modal {...args} />
+const Template: ComponentStory<typeof Modal> = (args) => <Modal {...args} />;
 
-export const Standard = Template.bind({}); 
+export const Standard = Template.bind({});
 Standard.args = {
-    title: 'Title',
-    subtitle: 'And som explanation',
-    width: "300px",
-    buttons: {
-        primary: {
-            label: 'Primary Button',
-            variant: 'primary',
-        },
-        secondary: {
-            label: 'Secondary Button',
-            variant: 'secondary',
-        },
-        alignment: "left"          
+  title: 'Title',
+  subtitle: 'And som explanation',
+  width: '300px',
+  buttons: {
+    primary: {
+      label: 'Primary Button',
+      variant: 'primary',
     },
-}
+    secondary: {
+      label: 'Secondary Button',
+      variant: 'secondary',
+    },
+    alignment: 'left',
+  },
+};
 
-export const ModalWithError = Template.bind({}); 
+export const ModalWithError = Template.bind({});
 ModalWithError.args = {
-    title: 'Title',
-    subtitle: 'And som explanation',
-    width: "300px",
-    buttons: {
-        primary: {
-            label: 'Primary Button',
-            variant: 'primary',
-        },
-        secondary: {
-            label: 'Secondary Button',
-            variant: 'secondary',
-        },
-        alignment: "left"          
+  title: 'Title',
+  subtitle: 'And som explanation',
+  width: '300px',
+  buttons: {
+    primary: {
+      label: 'Primary Button',
+      variant: 'primary',
     },
-    errorMessage: 'Woops, something went wrong',
-}
-
-
-
-
+    secondary: {
+      label: 'Secondary Button',
+      variant: 'secondary',
+    },
+    alignment: 'left',
+  },
+  errorMessage: 'Woops, something went wrong',
+};
