+++ src/Components/Modal/index.ts
@@ -1,1 +1,1 @@
-export { Modal } from "./Modal";
+export { Modal } from './Modal';
