+++ src/Icons/CheckRounded.tsx
@@ -1,20 +1,22 @@
 import React, { ReactElement } from 'react';
 
-function CheckRounded(): ReactElement {
-    return (
-        <svg
-            xmlns='http://www.w3.org/2000/svg'
-            height='24px'
-            viewBox='0 0 24 24'
-            width='24px'
-            fill='#01459E'
-        >
-            <path d='M0 0h24v24H0V0z' fill='none' />
-            <path d='M9 16.17L5.53 12.7c-.39-.39-1.02-.39-1.41
+const CheckRounded = (): ReactElement => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      height="24px"
+      viewBox="0 0 24 24"
+      width="24px"
+      fill="#01459E"
+    >
+      <path d="M0 0h24v24H0V0z" fill="none" />
+      <path
+        d="M9 16.17L5.53 12.7c-.39-.39-1.02-.39-1.41
              0-.39.39-.39 1.02 0 1.41l4.18 4.18c.39.39 1.02.39
-             1.41 0L20.29 7.71c.39-.39.39-1.02 0-1.41-.39-.39-1.02-.39-1.41 0L9 16.17z' />
-        </svg>
-    );
-}
+             1.41 0L20.29 7.71c.39-.39.39-1.02 0-1.41-.39-.39-1.02-.39-1.41 0L9 16.17z"
+      />
+    </svg>
+  );
+};
 
 export { CheckRounded };
