+++ src/Styles/Components/FileInputStyles.ts
@@ -1,36 +1,35 @@
-
 const FileInputStyles = {
-    baseStyle: {
-        padding:'16px',
-        border:'dashed 1px #01459E',
-        backgroundColor: '#DFEEF9',
-        color: '#4A4A4A',
-        _hover: {
-            border:'dashed 2px #002855'
-        }
+  baseStyle: {
+    padding: '16px',
+    border: 'dashed 1px #01459E',
+    backgroundColor: '#DFEEF9',
+    color: '#4A4A4A',
+    _hover: {
+      border: 'dashed 2px #002855',
     },
-    sizes: {
-        sm: {
-            width: '240px',
-            paddingLeft: '32px'
-        },
-        md: {
-            width:'392px'
-         }
+  },
+  sizes: {
+    sm: {
+      width: '240px',
+      paddingLeft: '32px',
     },
-    variants: {
-        disabled: {
-            backgroundColor: '#A4A4A4',
-            color: '#FFFFFF',
-            cursor: 'default',
-            _hover: {
-            border: 'dashed 1px #01459E'
-        }
-        }
+    md: {
+      width: '392px',
     },
-    defaultProps: {
-        size: 'md',
-    }
+  },
+  variants: {
+    disabled: {
+      backgroundColor: '#A4A4A4',
+      color: '#FFFFFF',
+      cursor: 'default',
+      _hover: {
+        border: 'dashed 1px #01459E',
+      },
+    },
+  },
+  defaultProps: {
+    size: 'md',
+  },
 };
 
 export { FileInputStyles };
