+++ src/Styles/Components/TableStyles.ts
@@ -1,14 +1,14 @@
 const TableStyles = {
-    variants: {
-        'simple': {
-            th: {
-                color: '#4A4A4A',
-                fontWeight: '700',
-                fontSize: '1rem',
-                textTransform:'none'
-            }
-        }
-    }
+  variants: {
+    'simple': {
+      th: {
+        color: '#4A4A4A',
+        fontWeight: '700',
+        fontSize: '1rem',
+        textTransform: 'none',
+      },
+    },
+  },
 };
 
 export { TableStyles };
